first commit so andy and I can git on the same page
main takeaway here is that the HTML will contain everything, and we will give classes to every div group so that we can change what displays on each page. it's fast, and should the JS fail (it won't but fail-safes like this are good standard practice), everything that someone might need from the site will still be available.